+++ Libraries/3rd/cairo/source/src/cairo-hash.c
@@ -204,7 +204,7 @@ _cairo_hash_table_create (cairo_hash_keys_equal_func_t keys_equal)
  * _cairo_hash_table_destroy is called. It is a fatal error otherwise,
  * and this function will halt. The rationale for this behavior is to
  * avoid memory leaks and to avoid needless complication of the API
- * with destroy notifiy callbacks.
+ * with destroy notify callbacks.
  *
  * WARNING: The hash_table must have no running iterators in it when
  * _cairo_hash_table_destroy is called. It is a fatal error otherwise,
@@ -340,7 +340,7 @@ _cairo_hash_table_lookup (cairo_hash_table_t *hash_table,
 {
     cairo_hash_entry_t *entry;
     unsigned long table_size, i, idx, step;
-    unsigned long hash = key->hash;
+    uintptr_t hash = key->hash;
 
     entry = hash_table->cache[hash & 31];
     if (entry && entry->hash == hash && hash_table->keys_equal (key, entry))
