+++ Libraries/3rd/cairo/source/src/cairo-xlib-source.c
@@ -487,7 +487,7 @@ color_source (cairo_xlib_surface_t *dst, const cairo_color_t *color)
 	    }
 
 	    gcv.foreground = 0;
-	    gcv.foreground |= color->alpha_short >> 8 << 24;
+	    gcv.foreground |= (uint32_t)color->alpha_short >> 8 << 24;
 	    gcv.foreground |= color->red_short   >> 8 << 16;
 	    gcv.foreground |= color->green_short >> 8 << 8;
 	    gcv.foreground |= color->blue_short  >> 8 << 0;
@@ -567,7 +567,7 @@ transparent_source (cairo_xlib_surface_t *dst, const cairo_color_t *color)
 {
     cairo_xlib_display_t *display = dst->display;
     uint32_t pixel =
-	color->alpha_short >> 8 << 24 |
+	(uint32_t)color->alpha_short >> 8 << 24 |
 	color->red_short   >> 8 << 16 |
 	color->green_short >> 8 << 8 |
 	color->blue_short  >> 8 << 0;
